pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "indi-rec-ia-agent"
version = "1.0.0"
description = "Agente de IA para INDI MAYORISTA"
authors = [
    {name = "Olvis Quintana", email = "olvisquintana@bcp.com.pe"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "azure-ai-documentintelligence==1.0.2",
    "azure-core==1.35.0",
    "azure-functions==1.23.0",
    "azure-servicebus==7.14.2",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "distro==1.9.0",
    "et_xmlfile==2.0.0",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "isodate==0.7.2",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.26",
    "langchain-core==0.3.69",
    "langchain-openai==0.3.28",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.5.3",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk==0.1.73",
    "langsmith==0.4.7",
    "numpy==2.3.1",
    "openai==1.97.0",
    "openpyxl==3.1.5",
    "orjson==3.11.0",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "pyodbc==5.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "redis==6.2.0",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.41",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "xlrd==2.0.2",
    "xxhash==3.5.0",
    "zstandard==0.23.0"
]

[project.urls]
"Homepage" = "https://github.com/CIX-BCP/indi-rec-ia-agent"
"Bug Tracker" = "https://github.com/CIX-BCP/indi-rec-ia-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]
